fix(Footer): COM-3413 Footer performance
