+++ packages/core-design/src/components/Footer/DoubleBottom/Addresses/Time/index.ts
@@ -1,0 +1,1 @@
+export * from './Time'
